+++ extras/scion-telegram/internal/telegram/format_test.go
