+++ extras/scion-telegram/internal/telegram/register_test.go
