+++ extras/scion-telegram/internal/telegram/register_v2.go
